# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Feiticeiros e Maldições – RPG Character Sheet API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /techniques:
    get:
      operationId: listTechniques
      tags: [techniques]
      summary: List available cursed techniques from the books
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of techniques
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Technique"
    post:
      operationId: createTechnique
      tags: [techniques]
      summary: Create a custom technique
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechniqueInput"
      responses:
        "201":
          description: Technique created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technique"
  /techniques/{id}:
    get:
      operationId: getTechnique
      tags: [techniques]
      summary: Get technique details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Technique details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technique"
        "404":
          description: Not found
...
components:
  schemas:
...
    Technique:
      type: object
      required: [id, name, category, description, isCustom]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: string
        source:
          type: ["string", "null"]
        abilities:
          type: ["string", "null"]
        isCustom:
          type: boolean
    TechniqueInput:
      type: object
      required: [name, category, description]
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
        description:
          type: string
        source:
          type: string
        abilities:
          type: string
...